PlantStressDetection/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.2.0
scikit-image>=0.20.0
opencv-python>=4.7.0
pandas>=2.0.0
numpy>=1.23.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
scipy>=1.10.0
xgboost>=1.7.0
rich>=13.0.0
streamlit>=1.28.0
plotly>=5.18.0
albumentations>=1.3.0
Pillow>=9.0.0
